Server/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
firebase-admin==5.0.0
python-dotenv==0.19.0
mysql-connector-python==8.0.26
scikit-learn==1.2.2
pandas
joblib
numpy


